requirements-edr.txt
psutil>=5.9.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=6.0
python-dateutil>=2.8.2
